requirements.txt
gradio==4.19.2
huggingface_hub>=0.20.0,<0.25.0
openai>=1.0.0
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
pydub>=0.25.1
python-dotenv>=1.0.0
numpy<2.0.0
ffmpeg-python>=0.2.0
